+++ project/fix_4.py
@@ -1,9 +1,14 @@
+import shutil
+
 
 class Fix_4_1:
 
-    def fix_container(container,fixes):
-        f = open('/tmp/Dockefile', mode='wt', encoding='utf-8')
-        f.write("FROM" + str(container.image) + "\n")
-        f.write("RUN useradd -d /home/username -m -s /bin/bash username" + "\n")
-        f.write("USER username" + "\n")
-        f.close()
+    def fix_container(container):
+        return 'RUN useradd -d /home/username -m -s /bin/bash username \nUSER username \n\n'
+
+
+class Fix_4_6:
+
+    def fix_container(container):
+        shutil.copyfile('../../templates/docker-healthcheck', '../../output/docker-healthcheck')
+        return 'COPY docker-healthcheck /usr/local/bin/ \nHEALTHCHECK CMD ["docker-healthcheck"]\n'