+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Proper README, CHANGELOG, CONTRIBUTING and LICENSE documentation.
 - Open api specification 1.0.0.
 - New endpoint to list all containers running on host
+- New endpoint to fix container
+- Healthcheck fix
 
 ### Security
 - Flask dependency updated to 1.0.0.